requirements.txt
annotated-types==0.7.0
dnspython==2.6.1
et-xmlfile==1.1.0
iniconfig==2.0.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
